refactor: remove 'util' from module name
