+++ packages/commons/src/util/ExponentialBackoff/index.ts
@@ -17,4 +17,4 @@
  *
  */
 
-export * from './ExponentialBackoffUtil';
+export * from './ExponentialBackoff';
